Add Media Saving & Convo Saving + Option To Not Delete Messages

+++ DiscordDMNuker/DiscordDMNuker/Main.cs
@@ -11,6 +11,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using Discord.Media;
+using Discord.Gateway;
+using System.Net.Mime;
+using System.Net;
 
 namespace DiscordDMNuker
 {
@@ -19,6 +23,8 @@ namespace DiscordDMNuker
         private ToolStripStatusLabel Status;
         private ListBox Logs;
         private int Index;
+
+        string currentPath = Directory.GetCurrentDirectory();
         public Main()
         {
             InitializeComponent();
@@ -26,18 +32,22 @@ namespace DiscordDMNuker
 
         private void Main_Load(object sender, EventArgs e)
         {
+            if (!Directory.Exists(Path.Combine(currentPath, "SavedMedia")))
+                Directory.CreateDirectory(Path.Combine(currentPath, "SavedMedia"));
 
+            if (!Directory.Exists(Path.Combine(currentPath, "SavedConvos")))
+                Directory.CreateDirectory(Path.Combine(currentPath, "SavedConvos"));
         }
 
-        private async void Start(string Token, ulong UserId, int Delay)
+        private async void Start(string Token, ulong UserId, int Delay, bool savepicsnvids, bool savemessages, bool delete)
         {
             await Task.Run(async () =>
            {
 
                try
                {
+                   Random rnd = new Random();
                    Status.SafeChangeText("Starting");
-
                    DiscordClient client = new DiscordClient(Token);
                    Logs.SafeAddItem(string.Format("Logged In To: {0}", client.User.Username));
                    var User = await client.GetProfileAsync(UserId);
@@ -45,19 +55,46 @@ namespace DiscordDMNuker
                    Logs.SafeAddItem(string.Format("Created Dms With: {0}", User.User.Username));
                    var msg = await client.GetChannelMessagesAsync(channelid.Id);
                    Status.SafeChangeText("In Progress....");
+                   var Convo = "SavedConvos/" + User.User.Username + rnd.Next(1, 999999999) + ".txt";
                    foreach (DiscordMessage message in msg)
                    {
-                       if (message.Author.User.Id == client.User.Id)
+                       if (savemessages)
                        {
-                           if (message.Type != MessageType.Call) {
-                               await message.DeleteAsync();
-                               Logs.SafeAddItem(string.Format("Deleted Message: {0}", message.Content));
-                               await Task.Delay(Delay);
+                           
+                           using (StreamWriter writetext = new StreamWriter(Convo, true))
+                           {
+                               writetext.WriteLine(message.Author.User.Username + " || " + message.Content);
+                           }
+                       }
+                       if (savepicsnvids)
+                       {
+                           if (message.Attachment != null)
+                           {
+                               if (message.Attachment.FileName.Contains(".mp4") || message.Attachment.FileName.Contains(".jpg") || message.Attachment.FileName.Contains(".png") || message.Attachment.FileName.Contains(".webm") || message.Attachment.FileName.Contains(".png") || message.Attachment.FileName.Contains(".gif"))
+                               {
+                                   var webClient = new WebClient();
+                                   webClient.DownloadFileCompleted += (sender, e) => Logs.SafeAddItem(string.Format("Saved Image/Video: {0}", message.Attachment.FileName)); ;
+                                   webClient.DownloadFileAsync(new Uri(message.Attachment.Url), "SavedMedia/" + message.Attachment.FileName);
+                               }
+                           }
+                       }
+                       if (delete)
+                       {
+                           if (message.Author.User.Id == client.User.Id)
+                           {
+                               if (message.Type != MessageType.Call)
+                               {
+                                   await message.DeleteAsync();
+                                   Logs.SafeAddItem(string.Format("Deleted Message: {0}", message.Content));
+                                   await Task.Delay(Delay);
+                               }
                            }
                        }
                    }
                    Status.SafeChangeText("Completed");
-               } catch (Exception) {
+               } catch (Exception ex) {
+                   //debug stuff
+                   Logs.SafeAddItem(string.Format("Error Message: {0}", ex.Message));
                    await Task.Delay(Delay);
                }
            });
@@ -74,7 +111,7 @@ namespace DiscordDMNuker
                     Logs = listBox1;
                     Index = 0;
 
-                    Start(FormStart.Token, FormStart.UserId, FormStart.Delay);
+                    Start(FormStart.Token, FormStart.UserId, FormStart.Delay, FormStart.SavePicsNVids, FormStart.SaveMessages, FormStart.Delete);
                 }
             }
         }
